ML/sentiment/nltk-transformer/nlp-analyzer/nlp-topic-analysis-pretrained-models/requirements.txt
--index-url https://pypi.org/simple

# Gradio UI
gradio>=4.0.0

# FastAPI REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mangum>=0.17.0

# Topic modeling
bertopic>=0.16.0
sentence-transformers>=2.2.2
umap-learn>=0.5.5
hdbscan>=0.8.33

# Classical topic modeling
scikit-learn>=1.3.0
gensim>=4.3.0
pyLDAvis>=3.4.0

# NLP
spacy>=3.7.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
nltk>=3.9.1
scipy>=1.11.0
numpy>=1.26.4

# Data / ML
pydantic>=2.9.0
pandas>=2.0.0
plotly>=5.3.1
matplotlib>=3.8.4

# Build tools
setuptools>=75.0.0
wheel>=0.44.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.1
